% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_det_prob.R
\name{calc_det_prob}
\alias{calc_det_prob}
\title{Calculate detection probabilities from data that were previously imported
with read_data()}
\usage{
calc_det_prob(
  data,
  selected_taxon_level = "scientificName",
  selected_taxon_id = "All",
  pool_primers = FALSE
)
}
\arguments{
\item{data}{(required, data.frame) Data.frame read in with \code{\link[=read_data]{read_data()}}.}
}
\value{
Two lists, each with distinct elements of the selected taxon; primer ID
containing 5-7 columns:
\itemize{
\item \code{month}
\item \code{n} total number of samples per month
\item \code{nd} number of positive detections per month
\item \code{p} calculated monthly detection probability
\item \code{s} standard deviation
\item \code{protocol_ID}
\item \code{year}
\item \code{yr.mo} year;month concatenated
}
}
\description{
Function that calculates non-parametric probability of detection
for each selected taxon and primer for the selected data. Probabilities are
calculated both (1) monthly, across all years; and (2) monthly with each year
separate. Outputs are used in subsequent functions.
}
\examples{
\dontrun{
 calc_det_prob(data = D_mb)
}
}
\author{
Anais Lacoursiere-Roussel \email{Anais.Lacoursiere@dfo-mpo.gc.ca}
}
