% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort_needed_fig.R
\name{effort_needed_fig}
\alias{effort_needed_fig}
\title{Calculate sampling effort needed to obtain different detection thresholds.}
\usage{
effort_needed_fig(
  scaledprobs,
  height_per_species = 400,
  dot_size = 12,
  selected_taxon_level = "scientificName"
)
}
\arguments{
\item{scaledprobs}{(required, data.frame) Normalized detection
probabilities as returned by the element \code{month} of the list returned by
\code{\link[=scale_newprob]{scale_newprob()}} for one taxon and one primer.}
}
\description{
This function calculates number of samples needed to obtain
the selected taxon detection at different thresholds by using scaled and interpolated
data produced with \verb{[scale_newprob()]}.
}
\examples{
\dontrun{
newprob <- calc_det_prob(gotedna_data$metabarcoding)
scaledprobs <- scale_newprob(gotedna_data$metabarcoding, newprob)
effort_needed_fig(
  scaledprobs
  )
)
}
}
\author{
Anais Lacoursiere-Roussel \email{Anais.Lacoursiere@dfo-mpo.gc.ca}
}
