% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_sample_fig.R
\name{field_sample_fig}
\alias{field_sample_fig}
\title{Display actual sampling effort of taxa within specified taxonomic group.}
\usage{
field_sample_fig(newprob, year, global_max_size, max_px = 1)
}
\arguments{
\item{data}{(required, data.frame): Data.frame read in with \code{\link[=read_data]{read_data()}}}
}
\description{
This function displays the actual number of eDNA samples taken
for all species of a specified taxonomic group per year.
}
\examples{
\dontrun{
field_sample_fig(
  data = gotedna_data$metabarcoding
)
}
}
\author{
Anais Lacoursiere-Roussel \email{Anais.Lacoursiere@dfo-mpo.gc.ca}
}
