% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higher_tax_fig.R
\name{higher_tax_fig}
\alias{higher_tax_fig}
\title{Display actual sampling effort of taxa within specified taxonomic group.}
\usage{
higher_tax_fig(
  data,
  higher.taxon.select,
  taxon.name,
  view.by.level,
  ecodistrict.select,
  primer.select
)
}
\arguments{
\item{data}{(required, data.frame): Data.frame read in with \code{\link[=read_data]{read_data()}}}

\item{higher.taxon.select}{(required, character): Higher taxonomic level of
interest. Choices = one of \code{c("kingdom", "phylum", "class", "order", "family")}}

\item{taxon.name}{(required, character): Select taxon name that matches the level
provided in \code{higher.taxon.select}. E.g., if \code{taxon.level = "class"}
enter class name, etc..}

\item{view.by.level}{(required, character): Lower taxonomic level to view
detection probability by. Cannot be higher level than higher group specified in
\code{higher.taxon.select} Choices = one of \code{c("phylum", "class", "order", "family", "genus")}}

\item{ecodistrict.select}{(required, character): Ecodistrict present in data.frame.}

\item{primer.select}{(required, character): Select primer as different primers
may provide different detection rates.}
}
\description{
This function displays the actual number of eDNA samples taken
for all taxa of a specified higher taxonomic group per year.
}
\examples{
\dontrun{
higher_tax_fig(data = D_mb, higher.taxon.select = "phylum", taxon.name = "Bryozoa",
view.by.level = "genus", ecodistrict.select = "Bay of Fundy", primer.select = "COI1")
}
}
\author{
Melissa Morrison \email{Melissa.Morrison@dfo-mpo.gc.ca}

Tim Barrett \email{Tim.Barrett@dfo-mpo.gc.ca}
}
