% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_fig.R
\name{hm_fig}
\alias{hm_fig}
\title{Displays heatmap of scaled monthly detection probabilities for taxon}
\usage{
hm_fig(scaledprobs, selected_taxon_level = "scientificName")
}
\arguments{
\item{scaledprobs}{(required, data.frame) Normalized detection
probabilities as returned by \code{\link[=scale_newprob]{scale_newprob()}}.}
}
\description{
This function displays scaled species detection probabilities
of the chosen taxonomic group with probabilities produced with
\code{\link[=scale_newprob]{scale_newprob()}}. All primers that are available in the data are
displayed.
NOTE: interpolated \verb{(i.e., missing)} data are not used in this
representation.
}
\examples{
\dontrun{
newprob <- calc_det_prob(D_mb)
scaledprobs <- scale_newprob(D_mb, newprob)
hm_fig(scaledprobs)
}
}
\author{
Anais Lacoursiere-Roussel \email{Anais.Lacoursiere@dfo-mpo.gc.ca}
}
