% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard_test.R
\name{jaccard_test}
\alias{jaccard_test}
\title{Calculate variation in optimal window among years}
\usage{
jaccard_test(scaledprobs, threshold)
}
\arguments{
\item{scaledprobs}{(required, list) Normalized detection probabilities
aggregated per month and year \code{\link[=scale_newprob]{scale_newprob()}}.}

\item{threshold}{(required, character): Detection probability threshold for
which data are to be displayed to visualize potential optimal detection
windows.
Choices = one of \code{c("50","55","60","65","70","75","80","85","90","95")}}
}
\value{
A data.frame with 10 columns:
\itemize{
\item \code{species}
\item \code{wt_mean} Weighted mean Jaccard Similarity Index.
\item \code{primer}
\item \code{kingdom}
\item \code{phylum}
\item \code{class}
\item \code{order}
\item \code{family}
\item \code{genus}
\item \code{GOTeDNA_ID.v}
}
}
\description{
This function calculates the Jaccard Similarity Index to provide
a measure of annual variation in species optimal eDNA detection window at the
desired detection threshold. The window is defined as consecutive months ≥
the threshold and executes a Fisher exact test comparing the detection
probability within and outside the window.
}
\examples{
\dontrun{
jaccard.test(data = D_mb, threshold = "75",
 taxon.name = "Acartia longiremis", scaledprobs)
}
}
\author{
Anais Lacoursiere-Roussel \email{Anais.Lacoursiere@dfo-mpo.gc.ca}
}
