% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer_sort.R
\name{primer_sort}
\alias{primer_sort}
\title{Sort primer choices by percent rank detection in data request window.}
\usage{
primer_sort(
  taxon.level = c("kingdom", "phylum", "class", "order", "family", "genus",
    "scientificName"),
  scaledprobs
)
}
\arguments{
\item{taxon.level}{taxonomic level.}

\item{scaledprobs_month}{(required, data.frame): Normalized detection.}
}
\description{
This function sorts the available primers by percent success
in detecting the selected taxa to allow selection of the optimal primer.
Ingests data produced with \verb{[scale_newprob()]}.
}
\examples{
\dontrun{
newprob <- calc_det_prob(gotedna_data$metabarcoding)
scaledprobs <- scale_newprob(gotedna_data$metabarcoding, newprob)
primer_sort("phylum", scaledprobs$Pscaled_month)
}
}
\author{
Anais Lacoursiere-Roussel \email{Anais.Lacoursiere@dfo-mpo.gc.ca}
}
