% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_bargraph.R
\name{protocol_bargraph}
\alias{protocol_bargraph}
\title{Run bargraph for protocol data and return interactive plot}
\usage{
protocol_bargraph(df)
}
\arguments{
\item{df}{data.frame with protocol info}
}
\value{
plotly object
}
\description{
Run bargraph for protocol data and return interactive plot
}
