% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmds_plot.R
\name{protocol_nmds}
\alias{protocol_nmds}
\title{Run NMDS on protocol data and return interactive plot}
\usage{
protocol_nmds(df)
}
\arguments{
\item{df}{data.frame with protocol info}
}
\value{
plotly object
}
\description{
Run NMDS on protocol data and return interactive plot
}
