% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_edna.R
\name{qpcr_fig}
\alias{qpcr_fig}
\title{qpcr_fig}
\usage{
qpcr_fig(
  data,
  manhattan = TRUE,
  heatmap = TRUE,
  ecoregion.select = NULL,
  taxon.select = NULL,
  species.select = NULL,
  year = FALSE,
  year.select = NULL,
  x.axis = "MONTH",
  title = NULL,
  free.scale = c("fixed", "free"),
  dashed.quartile = FALSE,
  dotted.mean = TRUE,
  color.palette = "viridis",
  height.override = NULL,
  width.override = NULL,
  filename.prefix = NULL,
  path.folder = NULL,
  write.image = FALSE
)
}
\arguments{
\item{data}{(object). Data generated by \code{\link{read_edna}}.}

\item{manhattan}{(optional, logical). Generate the manhattan plot.
Default: \code{manhattan = TRUE}.}

\item{heatmap}{(optional, logical). Generate the heatmap plot.
Default: \code{heatmap = TRUE}.}

\item{ecoregion.select}{(optional, character, string) If not supplied,
all the ecoregions in the data are used. To focus on fewer ecoregions,
list with this argument.
Default: \code{ecoregion.select = NULL}.}

\item{taxon.select}{(optional, character) Select one of the taxon.
Available values: \code{taxon.select = c("KINGDOM", "PHYLUM", "CLASS", "ORDER",
"FAMILY", "GENUS", "SCIENTIFIC_NAME", "SPP")}. With default the visualization
is generated for all taxon combined.
Default: \code{taxon.select = NULL}.}

\item{species.select}{(optional, character, string) If not supplied,
all the species available in the data are used. To focus on fewer species,
select and list the species with this argument. The name must be taken from
the column \code{SPP}, e.g. H. americanus.
Default: \code{species.select = NULL}.}

\item{year}{(optional, logical) To display the variable year in the figure.
Default: \code{year = FALSE}.}

\item{year.select}{(optional, integer, string) Select specific years to
display in the figure. By Default, all years available will be used, conditional
to \code{year = TRUE} is enabled.
Default: \code{year.select = NULL}.}

\item{x.axis}{(optional, character) By default the x axis display the variable month.
Using \code{x.axis = "SPP"}, will trigger the function to display the facet grid
by species on the x-axis and ecoregions using all months combined.
Default: \code{x.axis = "MONTH"}.}

\item{title}{(optional, character) Title to print on top of figure.
By default no title is displayed.
e.g. \code{title = "Manhattan plot of qPCR data by months and ecoregions"}.
Default: \code{title = NULL}.}

\item{free.scale}{(required, character) Should the scale of be free or fixed.
Available values: \code{free.scale = c("fixed", "free")}.}

\item{dashed.quartile}{(optional, logical) Display the Q25 and Q75 on the manhattan plot.
Default: \code{dashed.quartile = FALSE}.}

\item{dotted.mean}{(optional, logical) Display the mean frequency of detection on the manhattan plot.
Default: \code{dotted.mean = TRUE}.}

\item{color.palette}{(optional, character) Change the default color palette.
Available options are: \code{"viridis"}, ggplot2 basic colors: \code{"basic"} and
the ColorBrewer scales using \code{"brewer"}.
Default: \code{color.palette = "viridis"}.}

\item{height.override}{(optional, integer) Override the height of the plot.
Default: \code{height.override = NULL}.}

\item{width.override}{(optional, integer) Override the width of the plot.
Default: \code{width.override = NULL}.}

\item{filename.prefix}{(optional, character) Filename prefix.
Default: \code{filename.prefix = NULL}.}

\item{path.folder}{(optional, character)
Default: \code{path.folder = NULL}. Default will use the working directory.}

\item{write.image}{(optional, logical) Write the figure in \code{path.folder}.
Default: \code{write.image = FALSE}.}
}
\description{
The function generates a Manhattan and/or Heatmap figure
for frequency eDNA data
}
