% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{\emph{New proposal of this page}
Read and format metabarcoding metadata and data from OBIS}
\usage{
read_data(
  dataset_ids = NULL,
  scientificname = NULL,
  worms_id = NULL,
  areaid = NULL,
  join_by = c("auto", "occurrenceID", "id"),
  require_absences = TRUE
)
}
\value{
A tibble with 25 columns:
\itemize{
\item \code{protocol_ID}
\item \code{protocolVersion}
\item \code{samp_name}
\item \code{eventID}
\item \code{primer}
\item \code{species}
\item \code{domain}
\item \code{kingdom}
\item \code{phylum}
\item \code{class}
\item \code{order}
\item \code{family}
\item \code{genus}
\item \code{concentration}: provided when choose.method = "qPCR"
\item \code{pcr_primer_lod} : provided when choose.method = "qPCR"
\item \code{organismQuantity}: provided when choose.method = "metabarcoding"
\item \code{date}
\item ecodistrict`
\item \code{LClabel} : Local Contexts label to denote First Nations data sovereignty
\item \code{decimalLatitude}
\item \code{decimalLongitude}
\item \code{station}
\item \code{year}
\item \code{month}
\item \code{detected}
\item \code{msct} :logical, where minimum sequence copy threshold = 10
\item \code{ownerContact} : email of data owner/steward
\item \code{bibliographicCitation} : DOI reference, if applicable
}
}
\description{
Function that reads for any data in OBIS that has a DNA Derived Data
extension. Will compile occurrence (core) and DNA Derived Data (extension) files
into a single dataframe. Data frames are then formatted appropriately for use in
subsequence analysis and visualizations (e.g., date reformatting, merging metadata
and data).
\itemize{
\item Note that template column names are in the format of Darwin Core Archive
(DwC-A) using Darwin Core (DwC) data standards where possible.
}
}
\examples{
\dontrun{
D_mb <- read_data(
 choose.method = "metabarcoding", path.folder = "./inst/testdata"
)
}
}
\author{
Anais Lacoursiere-Roussel \email{Anais.Lacoursiere@dfo-mpo.gc.ca}
}
