% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_edna.R
\name{read_edna}
\alias{read_edna}
\title{read_edna}
\usage{
read_edna(path.folder = NULL, qpcr = TRUE)
}
\arguments{
\item{path.folder}{(optional, character)
Default: \code{path.folder = NULL}. Default will use the working directory.}

\item{qpcr}{(optional, logical) Read qPCR data.
Default: \code{qpcr = TRUE}.}
}
\value{
A tidy dataframe (tibble) with 15 columns:
\itemize{
\item\code{PROJECT_ID},
\item\code{SAMPLE_ID},
\item\code{DATA_QPCR_FILES}: added internally to keep track of files origins,
\item\code{METADATA_QPCR_FILES}: added internally to keep track of files origins,
\item\code{ECOREGION},
\item\code{KINGDOM},
\item\code{PHYLUM},
\item\code{CLASS},
\item\code{ORDER},
\item\code{FAMILY},
\item\code{GENUS},
\item\code{SCIENTIFIC_NAME}: e.g.: Homarus americanus,
\item\code{SPP}: e.g.: H. americanus,
\item\code{YEAR},
\item\code{MONTH},
\item\code{DETECTED}
}
}
\description{
Function that read several MS Excel files found in a specified folder.
When \code{qpcr = TRUE}, the function will search for MS Excel files with
\code{qPCR_data} in the name. Then it reads the metadata (sheet 2) and the
samples (sheet 3). Merge it together based on \code{EVENT_ID}.
}
\details{
The function looks for MS Excel files with \code{qPCR_data} in the filename.
In sheet 2 (the metadata information), the required and read columns are:
\code{"projectID", "sampleID", "eventID", "ecoregion", "samplingSite", "eventDate"}.
In sheet 2 (the samples and results), the required and read columns are:
\code{"eventID", "scientificName", "kingdom", "phylum", "class", "order",
"family", "genus", "occurrenceStatus"}.
}
\examples{
\dontrun{
data <- read_edna(path.folder = "optimal_eDNA_data", qpcr = TRUE)
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com} and Melissa Morrison \email{Melissa.Morrison@dfo-mpo.gc.ca}
}
