% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_fig.R
\name{sample_size_fig}
\alias{sample_size_fig}
\title{Display actual sampling effort by species, year, and region.}
\usage{
sample_size_fig(data, species.name, ecodistrict.select)
}
\arguments{
\item{data}{(required, data.frame): Data.frame read in with \code{read_data()}.}

\item{species.name}{(required, character): Full binomial species name.}

\item{ecodistrict.select}{(required, character): Ecodistrict present in data.frame.}
}
\description{
This function displays the actual number of eDNA samples taken
for the specified species and region.
}
\examples{
\dontrun{
sample_size_fig(data = D_mb, species.name = "Acartia longiremis",
ecodistrict.select = "Bay of Fundy")
}
}
\author{
Melissa Morrison \email{Melissa.Morrison@dfo-mpo.gc.ca}

Tim Barrett \email{Tim.Barrett@dfo-mpo.gc.ca}
}
