% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_newprob.R
\name{scale_newprob}
\alias{scale_newprob}
\title{Normalize detection probabilities with min-max method (range 011) by month
for each year separately.}
\usage{
scale_newprob(data, newprob, selected_taxon_level = "scientificName")
}
\arguments{
\item{data}{(required, data.frame) Data.frame imported with \code{\link[=read_data]{read_data()}}. Required
to join taxonomic information.}

\item{newprob}{(required, list) detection probabilities aggregated per month and year
\code{\link[=calc_det_prob]{calc_det_prob()}}.}
}
\value{
Grouped data.frame with 15 columns (though lower taxonomic levels are not included beyond the selected taxonomic level):
\itemize{
\item \code{id}: unique taxon;primer;year identifier
\item \code{month}:
\item \code{detect}: number of detections
\item \code{nondetect}: number of non-detections
\item \code{scaleP}: detection probability scaled to range 0-1
\item \code{protocol_ID}:
\item \code{species}:
\item \code{primer}:
\item \code{year}:
\item \code{phylum}:
\item \code{class}:
\item \code{order}:
\item \code{family}:
\item \code{genus}:
\item \code{fill}:
}
}
\description{
This function normalizes (i.e., scales) monthly detection
probabilities for each selected taxon (e.g. species), primer, and year that were calculated with
the previous function, \code{\link[=calc_det_prob]{calc_det_prob()}}. Outputs fed into figure and window
calculation functions.
\itemize{
\item NOTE: Currently this function only works for metabarcoding data.
}
}
\examples{
\dontrun{
newprob <- calc_det_prob(gotedna_data$metabarcoding, "Scotian Shelf")
scale_newprob(data = gotedna_data$metabarcoding, newprob)
}
}
\author{
Anais Lacoursiere-Roussel \email{Anais.Lacoursiere@dfo-mpo.gc.ca}
}
