% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_prob_by_month.R
\name{scale_prob_by_month}
\alias{scale_prob_by_month}
\title{Normalize detection probabilities with min-max method (range 0-1) by year and month}
\usage{
scale_prob_by_month(data, ecodistrict.select)
}
\arguments{
\item{data}{(required, data.frame) Data.frame imported with read_data(). Required
to join taxonomic information.}

\item{ecodistrict.select}{(required, character) Ecodistrict present in data.frame.}
}
\value{
Grouped data.frame with 15 columns:
\itemize{
\item\code{id}: unique species;primer identifier
\item\code{ecodistrict},
\item\code{month},
\item\code{detect}: number of detections
\item\code{nondetect}: number of non-detections
\item\code{scaleP}:  detection probability scaled to range \link{0,1}
\item\code{GOTeDNA_ID},
\item\code{species},
\item\code{primer},
\item\code{phylum},
\item\code{class},
\item\code{order},
\item\code{family},
\item\code{genus},
\item\code{fill},
}
}
\description{
This function normalizes (i.e., scales) monthly detection probabilities for
each species and primer that were calculated with the previous function,
\code{calc_det_prob()}. Outputs fed into figure and window calculation functions.
}
\examples{
\dontrun{
scale_prob_by_month(data = D_mb, ecodistrict.select = "Scotian Shelf")
}
}
\author{
Tim Barrett \email{Tim.Barrett@dfo-mpo.gc.ca}

Melissa Morrison \email{Melissa.Morrison@dfo-mpo.gc.ca}
}
