% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_prob_by_year.R
\name{scale_prob_by_year}
\alias{scale_prob_by_year}
\title{Normalize detection probabilities with min-max method (range 011) by month
for each year separately.}
\usage{
scale_prob_by_year(data, ecodistrict.select)
}
\arguments{
\item{data}{(required, data.frame) Data.frame imported with \code{\link[=read_data]{read_data()}}. Required
to join taxonomic information.}

\item{ecodistrict.select}{(required, character) Ecodistrict present in data.frame.}
}
\value{
Grouped data.frame with 16 columns:
\itemize{
\item \code{id} unique species;primer;year identifier
\item \code{ecodistrict}
\item \code{month}
\item \code{detect} number of detections
\item \code{nondetect} number of non-detections
\item \code{scaleP} detection probability scaled to range 0-1
\item \code{GOTeDNA_ID}
\item \code{species}
\item \code{primer}
\item \code{year}
\item \code{phylum}
\item \code{class}
\item \code{order}
\item \code{family}
\item \code{genus}
\item \code{fill}
}
}
\description{
This function normalizes (i.e., scales) monthly detection
probabilities for each species, primer, and year that were calculated with
the previous function, \code{\link[=calc_det_prob]{calc_det_prob()}}. Outputs fed into figure and window
calculation functions.
\itemize{
\item NOTE: Currently this function only works for metabarcoding data.
}
}
\examples{
\dontrun{
scale_prob_by_year(data = D_mb, ecodistrict.select = "Scotian Shelf")
}
}
\author{
Tim Barrett \email{Tim.Barrett@dfo-mpo.gc.ca}

Melissa Morrison \email{Melissa.Morrison@dfo-mpo.gc.ca}
}
