% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scientific_name_formatter}
\alias{scientific_name_formatter}
\title{Display genus and species names in italics}
\usage{
scientific_name_formatter(raw_name)
}
\arguments{
\item{raw_name}{Raw name.}
}
\description{
Display genus and species names in italics
}
\keyword{internal}
