% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_fig.R
\name{smooth_fig}
\alias{smooth_fig}
\title{Display species monthly detection probability by primer and region.}
\usage{
smooth_fig(scaledprobs)
}
\arguments{
\item{data}{(required, data.frame): Data.frame read in with \code{\link[=read_data]{read_data()}}.}
}
\description{
This function displays the species optimal detection period as
monthly detection probabilities with LOESS smoothing over years.
}
\examples{
\dontrun{
data <- gotedna_data$metabarcoding |> dplyr::filter(
     species == "Acartia longiremis",
     primer == "COI1"
 )
smooth_fig(data = data)
}
}
\author{
Anais Lacoursiere-Roussel \email{Anais.Lacoursiere@dfo-mpo.gc.ca}
}
