% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_window.R
\name{calc_window}
\alias{calc_window}
\title{Calculate species optimal detection window}
\usage{
calc_window(data, threshold, species.name, scaledprobs)
}
\arguments{
\item{data}{(required, data.frame) Data.frame read in with \code{\link[=read_data]{read_data()}}.}

\item{threshold}{(required, character): Detection probability threshold for
which data are to be displayed to visualize potential optimal detection
windows.
Choices = one of \code{c("50","55","60","65","70","75","80","85","90","95")}}

\item{species.name}{(required, character): Full binomial species name.}

\item{scaledprobs}{(required, list) Normalized detection probabilities
aggregated per month and year \code{\link[=scale_newprob]{scale_newprob()}}.}
}
\value{
A data.frame with 16 columns:
\itemize{
\item \code{length} Number of months with optimal detection \verb{(i.e., over the specified threshold)}.
\item \code{threshold} Detection probability threshold specified in function call.
\item \code{period} Range of months having optimal detection.
\item \code{GOTeDNA_ID}
\item \code{species}
\item \code{primer}
\item \code{phylum}
\item \code{class}
\item \code{order}
\item \code{family}
\item \code{genus}
\item \verb{odds ratio}
\item \verb{p value}
\item \verb{Lower CI}
\item \verb{Upper CI}
\item \code{confidence} Provided such that p value < 0.001 = Very high;
0.001 < p < 0.01 = High; 0.01 < p < 0.05 = Medium; p >= 0.05 = Low.
}
}
\description{
This function calculates species optimal eDNA detection period
\verb{(i.e., months)} and window \verb{(i.e., length/number of months)} at the
desired detection threshold. The window is defined as consecutive months ≥
the threshold and executes a Fisher exact test comparing the detection
probability within and outside the window. Output provides the odds ratio,
p-value, and confidence interval for each primer of the species selected,
showing the variation within year and among years, if applicable
\verb{(i.e., input = Pscaled_agg and Pscaled_yr)}.
}
\examples{
\dontrun{
newprob <- calc_det_prob(
  data = D_mb_ex,
  ecodistrict.select = "Scotian Shelf"
)
scaledprobs <- scale_newprob(
  D_mb_ex, "Scotian Shelf",
  newprob
)
Pscaled_month <- scale_prob_by_month(
  D_mb_ex, "Scotian Shelf",
  newprob$newP_agg
)
calc_window(
  data = D_mb_ex, threshold = "90",
  species.name = "Nucula proxima", scaledprobs
)
}
}
\author{
Melissa Morrison \email{Melissa.Morrison@dfo-mpo.gc.ca}

Tim Barrett \email{Tim.Barrett@dfo-mpo.gc.ca}
}
