% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort_needed_fig.R
\name{effort_needed_fig}
\alias{effort_needed_fig}
\title{Calculate sampling effort needed to obtain different detection thresholds.}
\usage{
effort_needed_fig(species.name, primer.select, ecodistrict.select)
}
\arguments{
\item{species.name}{(required, character): Full binomial species name.}

\item{primer.select}{(required, character): Select primer as different primers
may provide different detection rates.}

\item{ecodistrict.select}{(required, character): Ecodistrict present in data.frame.}
}
\description{
This function calculates number of samples needed to obtain
species detection at different thresholds by using scaled and interpolated
data produced with \verb{[scale_prob_by_month()]}
}
\examples{
\dontrun{
effort_needed_fig(
  species.name = "Acartia hudsonica", primer.select = "COI1",
  ecodistrict.select = "Scotian Shelf"
)
}
}
\author{
Melissa Morrison \email{Melissa.Morrison@dfo-mpo.gc.ca}

Tim Barrett \email{Tim.Barrett@dfo-mpo.gc.ca}
}
