% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort_needed_fig.R
\name{effort_needed_fig}
\alias{effort_needed_fig}
\title{Calculate sampling effort needed to obtain different detection thresholds.}
\usage{
effort_needed_fig(species.name, primer.select, scaledprobs)
}
\arguments{
\item{species.name}{(required, character): Full binomial species name.}

\item{primer.select}{(required, character): Select primer as different
primers may provide different detection rates.}

\item{scaledprobs}{(required, data.frame) Normalized detection
probabilities as returned by \code{\link[=scale_newprob]{scale_newprob()}}.}
}
\description{
This function calculates number of samples needed to obtain
species detection at different thresholds by using scaled and interpolated
data produced with \verb{[scale_newprob()]}.
}
\examples{
\dontrun{
newprob <- calc_det_prob(D_mb_ex)
scaledprobs <- scale_newprob(D_mb_ex, newprob)
effort_needed_fig(
  species.name = "Acartia hudsonica", primer.select = "COI1",
  scaledprobs)
}
}
\author{
Melissa Morrison \email{Melissa.Morrison@dfo-mpo.gc.ca}

Tim Barrett \email{Tim.Barrett@dfo-mpo.gc.ca}
}
