% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort_needed_fig.R
\name{effort_needed_fig}
\alias{effort_needed_fig}
\title{Calculate sampling effort needed to obtain different detection thresholds.}
\usage{
effort_needed_fig(scaledprobs_month)
}
\arguments{
\item{scaledprobs_month}{(required, data.frame) Normalized detection
probabilities as returned by the element \code{month} of the list returned by
\code{\link[=scale_newprob]{scale_newprob()}} for one species ans one primer.}
}
\description{
This function calculates number of samples needed to obtain
species detection at different thresholds by using scaled and interpolated
data produced with \verb{[scale_newprob()]}.
}
\examples{
\dontrun{
newprob <- calc_det_prob(D_mb_ex)
scaledprobs <- scale_newprob(D_mb_ex, newprob)
effort_needed_fig(
  scaledprobs$Pscaled_month |> dplyr::filter(
      species \%in\% "Acartia hudsonica",
      primer == "COI1"
  )
)
}
}
