% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_sample_fig.R
\name{field_sample_fig}
\alias{field_sample_fig}
\title{Display actual sampling effort of taxa within specified taxonomic group.}
\usage{
field_sample_fig(data, taxon.select, taxon.name)
}
\arguments{
\item{data}{(required, data.frame): Data.frame read in with \code{\link[=read_data]{read_data()}}}

\item{taxon.select}{(required, character): Taxonomic level of interest
Choices = one of \verb{c("kingdom", "phylum", "class", "order", "family", "genus",}
\verb{"species")}.}

\item{taxon.name}{(required, character): Select taxon name that matches the level
provided in \code{taxon.select}. E.g., if \code{taxon.level = "class"}
enter class name, etc..}
}
\description{
This function displays the actual number of eDNA samples taken
for all species of a specified taxonomic group per year.
}
\examples{
\dontrun{
higher_tax_fig(
  data = D_mb_ex,
  higher.taxon.select = "phylum",
  taxon.name = "Bryozoa"
)
}
}
\author{
Anais Lacoursiere-Roussel \email{Anais.Lacoursiere@dfo-mpo.gc.ca}
}
