% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_fig.R
\name{hm_fig}
\alias{hm_fig}
\title{Displays heatmap of scaled monthly detection probabilities for taxon}
\usage{
hm_fig(taxon.level, taxon.name, ecodistrict.select)
}
\arguments{
\item{taxon.level}{(required, data.frame): Select taxonomic level to view.
Choices = one of \code{c("phylum", "class", "order", "family", "genus", "species")}}

\item{taxon.name}{(required, character): Select taxon name that matches the level
provided in \code{taxon.level}. E.g., if \code{taxon.level = "genus"} enter genus name, etc.}

\item{ecodistrict.select}{(required, character): Ecodistrict present in data.frame.}
}
\description{
This function displays scaled species detection probabilities of the
chosen taxonomic group with probabilities produced with \code{scale_prob_by_month()}
All primers that are available in the data are displayed.
NOTE: interpolated \verb{(i.e., missing)} data are not used in this representation.
}
\examples{
\dontrun{
hm_fig(taxon.level = "class", taxon.name = "Copepoda",
ecodistrict.select = "Bay of Fundy")
}
}
\author{
Melissa Morrison \email{Melissa.Morrison@dfo-mpo.gc.ca}

Tim Barrett \email{Tim.Barrett@dfo-mpo.gc.ca}
}
