% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read and format qPCR and metabarcoding metadata and data from GOTeDNA templates}
\usage{
read_data(choose.method = c("qPCR", "metabarcoding"), path.folder = NULL)
}
\arguments{
\item{choose.method}{(required, character) Choices = c("qPCR", "metabarcoding").}

\item{path.folder}{(optional, character)
Default: \code{path.folder = NULL}. Default will use the current working directory.}
}
\value{
A tibble with 26 columns:
\itemize{
\item \code{GOTeDNA_ID}
\item \code{GOTeDNA_version}
\item \code{materialSampleID}
\item \code{eventID}
\item \code{primer}
\item \code{species}
\item \code{domain}
\item \code{kingdom}
\item \code{phylum}
\item \code{class}
\item \code{order}
\item \code{family}
\item \code{genus}
\item \code{concentration}: provided when choose.method = "qPCR"
\item \code{pcr_primer_lod} : provided when choose.method = "qPCR"
\item \code{organismQuantity}: provided when choose.method = "metabarcoding"
\item \code{date}
\item \code{ecodistrict}
\item \code{LClabel} : Local Contexts label to denote First Nations data sovereignty
\item \code{decimalLatitude}
\item \code{decimalLongitude}
\item \code{station}
\item \code{year}
\item \code{month}
\item \code{detected}
\item \code{msct} :logical, where minimum sequence copy threshold = 10
\item \code{ownerContact} : email of data owner/steward
\item \code{bibliographicCitation} : DOI reference, if applicable
}
}
\description{
Function that reads GOTeDNA qPCR or metabarcoding template MS
Excel sheets found in a specified folder. When \verb{choose.methods = "qPCR" or "metabarcoding"} it will compile metadata (sheet 2) and data (sheet 3 or 4,
respectively) into a single dataframe. Data frames are then formatted
appropriately for use in subsequent analysis and visualizations (e.g., date
reformatting, merging metadata and data).
\itemize{
\item Note that template column names are in the format of Darwin Core Archive
(DwC-A) using Darwin Core (DwC) data standards where possible.
}
}
\examples{
\dontrun{
D_mb <- read_data(
 choose.method = "metabarcoding", path.folder = "./inst/testdata"
)
}
}
\author{
Anais Lacoursiere-Roussel \email{Anais.Lacoursiere@dfo-mpo.gc.ca}
}
