% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_newprob.R
\name{scale_newprob}
\alias{scale_newprob}
\title{Normalize detection probabilities with min-max method (range 011) by month
for each year separately.}
\usage{
scale_newprob(data, ecodistrict.select, newprob)
}
\arguments{
\item{data}{(required, data.frame) Data.frame imported with \code{\link[=read_data]{read_data()}}. Required
to join taxonomic information.}

\item{ecodistrict.select}{(required, character) Ecodistrict present in data.frame.}

\item{newprob}{(required, list) detection probabilities aggregated per month and year
\code{\link[=calc_det_prob]{calc_det_prob()}}.}
}
\value{
Grouped data.frame with 16 columns:
\itemize{
\item \code{id} unique species;primer;year identifier
\item \code{ecodistrict}
\item \code{month}
\item \code{detect} number of detections
\item \code{nondetect} number of non-detections
\item \code{scaleP} detection probability scaled to range 0-1
\item \code{GOTeDNA_ID}
\item \code{species}
\item \code{primer}
\item \code{year}
\item \code{phylum}
\item \code{class}
\item \code{order}
\item \code{family}
\item \code{genus}
\item \code{fill}
}
}
\description{
This function normalizes (i.e., scales) monthly detection
probabilities for each species, primer, and year that were calculated with
the previous function, \code{\link[=calc_det_prob]{calc_det_prob()}}. Outputs fed into figure and window
calculation functions.
\itemize{
\item NOTE: Currently this function only works for metabarcoding data.
}
}
\examples{
\dontrun{
newprob <- calc_det_prob(D_mb_ex, "Scotian Shelf")
scale_newprob(data = D_mb, ecodistrict.select = "Scotian Shelf", newprob)
}
}
