% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_fig.R
\name{smooth_fig}
\alias{smooth_fig}
\title{Display species monthly detection probability by primer and region.}
\usage{
smooth_fig(data, species.name, primer.select)
}
\arguments{
\item{data}{(required, data.frame): Data.frame read in with \code{\link[=read_data]{read_data()}}.}

\item{species.name}{(required, character): Full binomial species name.}

\item{primer.select}{(required, character): Select primer as different
primers may provide different detection rates.}
}
\description{
This function displays the species optimal detection period as
monthly detection probabilities with LOESS smoothing over years.
}
\examples{
\dontrun{
smooth_fig(
  data = D_mb_ex, species.name = "Acartia longiremis", 
  primer.select = "COI1"
)
}
}
