% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh_fig.R
\name{thresh_fig}
\alias{thresh_fig}
\title{Display monthly detection probabilities for selected taxon, and
detection probability threshold}
\usage{
thresh_fig(taxon.level, taxon.name, threshold, scaledprobs)
}
\arguments{
\item{taxon.level}{(required, data.frame): Select taxonomic level to view.
Choices = one of \code{c("phylum", "class", "order", "family", "genus", "species")}}

\item{taxon.name}{(required, character): Select taxon name that matches the level
provided in \code{taxon.level}. E.g., if \code{taxon.level = "genus"} enter genus name, etc.}

\item{threshold}{(required, character): Detection probability threshold for
which data are to be displayed to visualize potential optimal detection windows.
Choices = one of \verb{"50","55","60","65","70","75","80","85","90","95")}}

\item{scaledprobs}{(required, data.frame) Normalized detection
probabilities as returned by \code{\link[=scale_newprob]{scale_newprob()}}.}
}
\description{
This function displays detection probabilities in bins, with
probabilities over a specified threshold displayed in purple and under the
threshold in grey.
}
\examples{
\dontrun{
newprob <- calc_det_prob(D_mb_ex)
scaledprobs <- scale_newprob(D_mb_ex, newprob)
thresh_fig(
  taxon.level = "species", taxon.name = "Acartia hudsonica",
  threshold = "90", scaledprobs
)
}
}
\author{
Anais Lacoursiere-Roussel \email{Anais.Lacoursiere@dfo-mpo.gc.ca}
}
